reorganized backend routes to align with our documentation – user and auth are now separate from api
